pyproject.toml
[project]
name = "afa-benchmark"
version = "0.4.0"
description = "A benchmark of active feature acquisition (AFA) methods."
readme = "README.md"
authors = [
    { name = "Valter Schutz", email = "valterschutz@proton.me" },
    { name = "Han Wu", email = "hanwu@chalmers.se" },
    { name = "Reza Rezvan", email = "rezvan@chalmers.se" },
    { name = "Linus Aronsson", email = "linaro@chalmers.se" },
]
requires-python = "==3.12.10"
dependencies = [
    "certifi>=2025.4.26",
    "coolname>=2.2.0",
    "cupy-cuda12x>=13.5.1; sys_platform == 'linux'",
    "dacite>=1.9.2",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "jaxtyping>=0.3.2",
    "lightning>=2.5.1.post0",
    "matplotlib>=3.10.1",
    "omegaconf>=2.3.0",
    "pandas>=2.2.3",
    "plotly>=6.1.2",
    "pytest>=8.3.5",
    "rich>=14.0.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "timm>=1.0.22",
    "torchrl>=0.8.0",
    "torchvision>=0.22.0",
    "tornado>=6.5.1",
    "ucimlrepo>=0.0.7",
    "wandb>=0.19.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["afabench"]

[dependency-groups]
dev = [
    "ipython>=8.36.0",
    "pre-commit>=4.3.0",
    "ruff>=0.11.8",
    "snakemake>=9.12.0",
    "snakemake-executor-plugin-slurm>=1.8.0",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.mypy]
follow_untyped_imports = true
